requirements.txt
python-dotenv==1.0.0
requests>=2.31.0
supabase>=2.0.0
pykrx>=1.0.0
yfinance>=0.2.31
python-telegram-bot>=21.0
google-genai>=1.0.0
numpy>=1.26.0
scipy>=1.12.0
pytz==2024.1
pytest>=8.0.0
# Stock Insight (modules/news/)
feedparser>=6.0.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.100.0
youtube-transcript-api>=0.6.0
curl_cffi>=0.5.10
